refactor: `input` package 추가
input과 관련된 클래스들을 모두 `input` package로 이동
