+++ src/main/java/racingcar/input/InputParser.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.input;
 
 import java.util.Arrays;
 import java.util.List;
