+++ src/test/java/racingcar/input/KeyboardInputParserTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.input;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
